automatically backed up by learn

+++ db/migrate/001_create_costumes.rb
@@ -14,7 +14,7 @@ class CreateCostumes < ActiveRecord::Migration[4.2]
     create_table :costumes do |t|
       t.string :name
       t.string :price
-      t.decimal :size
+      t.integer :size
       t.binary :image_url
     end
   end
